+++ plugins/11_TheWarWithin/zones/undermine.lua
@@ -18,6 +18,8 @@ local Item = ns.reward.Item
 local Mount = ns.reward.Mount
 local Pet = ns.reward.Pet
 local Reputation = ns.reward.Reputation
+local Section = ns.reward.Section
+local Spacer = ns.reward.Spacer
 local Toy = ns.reward.Toy
 local Transmog = ns.reward.Transmog
 
@@ -407,6 +409,32 @@ map.nodes[34097126] = Vendor({
         Toy({item = 237345, count = 5}) -- Limited Edition Rocket Bobber
     }
 }) -- Blair Bass
+
+map.nodes[24606320] = Vendor({
+    id = 236849,
+    note = L['greexit_coarsebub_note'],
+    rewards = {
+        Section('{faction:2673}'), -- Bilgewater
+        Transmog({item = 231742, count = 50}), -- Bilgewater Bruisers Helm
+        Transmog({item = 231748, count = 25}), -- Bilgewater Bruisers Spaulders
+        Transmog({item = 231736, count = 125}), -- Bilgewater Bruisers Tabard
+        Spacer(), --
+        Section('{faction:2677}'), -- Steamwheedle
+        Transmog({item = 231740, count = 50}), -- Steamwheedle Bruisers Helm
+        Transmog({item = 231747, count = 25}), -- Steamwheedle Bruisers Spaulders
+        Transmog({item = 231735, count = 125}), -- Steamwheedle Bruisers Tabard
+        Spacer(), --
+        Section('{faction:2675}'), -- Blackwater
+        Transmog({item = 231741, count = 50}), -- Blackwater Bruisers Helm
+        Transmog({item = 231746, count = 25}), -- Blackwater Bruisers Spaulders
+        Transmog({item = 231734, count = 125}), -- Blackwater Bruisers Tabard
+        Spacer(), --
+        Section('{faction:2671}'), -- Venture Company
+        Transmog({item = 231744, count = 50}), -- Venture Co. Bruisers Helm
+        Transmog({item = 231750, count = 25}), -- Venture Co. Bruisers Spaulders
+        Transmog({item = 231738, count = 125}) -- Venture Co. Bruisers Tabard
+    }
+}) -- Greexit Coarsebub
 
 -------------------------------------------------------------------------------
 ----------------------------- S.C.R.A.P. REWARDS ------------------------------